git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@4935 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -1504,7 +1504,7 @@ class CVertColorData {
 
 typedef struct tFaceListItem {
 	CSegFace*			faceP;
-	short					nNextItem;
+	int					nNextItem;
 } tFaceListItem;
 
 class CFaceListIndex {
